apps/rag-eval/pyproject.toml
[project]
name = "rag-eval"
version = "0.3.0"
description = "RAG Evaluation CLI - Search and manage evaluation datasets"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "typer>=0.12.0",
    "rich>=13.0.0",
    "huggingface-hub>=1.3.0",
    "datasets>=3.0.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "respx>=0.22.0",
]

[project.scripts]
rag-eval = "rag_eval.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rag_eval"]
